Refactor crate into focused modules
