+++ src/table.rs
@@ -1,0 +1,168 @@
+//! Markdown table reflow utilities.
+//!
+//! Implements the algorithm outlined in `docs/html-table-support.md` lines 1-24.
+//! Provides helpers used by the `reflow` module and `reflow_table` itself.
+
+use regex::Regex;
+
+fn next_is_pipe(chars: &mut std::iter::Peekable<std::str::Chars<'_>>) -> bool {
+    chars.peek() == Some(&'|')
+}
+
+#[must_use]
+pub fn split_cells(line: &str) -> Vec<String> {
+    let mut s = line.trim();
+    if let Some(stripped) = s.strip_prefix('|') {
+        s = stripped;
+    }
+    if let Some(stripped) = s.strip_suffix('|') {
+        s = stripped;
+    }
+
+    let mut cells = Vec::new();
+    let mut current = String::new();
+    let mut chars = s.chars().peekable();
+    while let Some(ch) = chars.next() {
+        if ch == '\\' {
+            if next_is_pipe(&mut chars) {
+                chars.next();
+                current.push('|');
+                continue;
+            }
+            current.push(ch);
+            continue;
+        }
+        if ch == '|' {
+            cells.push(current.trim().to_string());
+            current.clear();
+        } else {
+            current.push(ch);
+        }
+    }
+    cells.push(current.trim().to_string());
+    cells
+}
+
+pub(crate) fn format_separator_cells(widths: &[usize], sep_cells: &[String]) -> Vec<String> {
+    if sep_cells.len() != widths.len() {
+        return sep_cells.to_vec();
+    }
+
+    sep_cells
+        .iter()
+        .enumerate()
+        .map(|(i, cell)| {
+            let trimmed = cell.trim();
+            let left = trimmed.starts_with(':');
+            let right = trimmed.ends_with(':');
+            let mut dashes = "-".repeat(widths[i].max(3));
+            if left {
+                dashes.remove(0);
+                dashes.insert(0, ':');
+            }
+            if right {
+                dashes.pop();
+                dashes.push(':');
+            }
+            dashes
+        })
+        .collect()
+}
+
+fn sep_index_within(idx: Option<usize>, len: usize) -> Option<usize> {
+    match idx {
+        Some(i) if i < len => Some(i),
+        _ => None,
+    }
+}
+
+fn rows_mismatched(rows: &[Vec<String>], split_within_line: bool) -> bool {
+    if split_within_line {
+        return false;
+    }
+    let Some(first_len) = rows.first().map(Vec::len) else {
+        return false;
+    };
+    rows.iter()
+        .skip(1)
+        .any(|row| row.len() != first_len && !row.iter().all(|c| SEP_RE.is_match(c)))
+}
+
+pub(crate) static SEP_RE: std::sync::LazyLock<Regex> =
+    std::sync::LazyLock::new(|| Regex::new(r"^[\s|:-]+$").unwrap());
+
+#[must_use]
+pub fn reflow_table(lines: &[String]) -> Vec<String> {
+    if lines.is_empty() {
+        return Vec::new();
+    }
+
+    let indent: String = lines[0].chars().take_while(|c| c.is_whitespace()).collect();
+    let mut trimmed: Vec<String> = lines
+        .iter()
+        .map(|l| l.trim().to_string())
+        .filter(|l| !l.trim_start().starts_with("\\-"))
+        .collect();
+    let sep_idx = trimmed.iter().position(|l| SEP_RE.is_match(l));
+    let sep_line = sep_idx.map(|idx| trimmed.remove(idx));
+
+    let (rows, split_within_line) = crate::reflow::parse_rows(&trimmed);
+
+    let max_cols = rows.iter().map(Vec::len).max().unwrap_or(0);
+
+    let (sep_cells, sep_row_idx) =
+        crate::reflow::detect_separator(sep_line.as_ref(), &rows, max_cols);
+
+    let cleaned = crate::reflow::clean_rows(rows);
+
+    let mut output_rows = cleaned.clone();
+    if let Some(idx) = sep_index_within(sep_row_idx, output_rows.len()) {
+        output_rows.remove(idx);
+    }
+
+    if rows_mismatched(&cleaned, split_within_line) {
+        return lines.to_vec();
+    }
+
+    let widths = crate::reflow::calculate_widths(&cleaned, max_cols);
+
+    let out = crate::reflow::format_rows(output_rows, &widths, &indent);
+
+    crate::reflow::insert_separator(out, sep_cells, &widths, &indent)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn sep_index_within_bounds() {
+        assert_eq!(sep_index_within(Some(1), 3), Some(1));
+        assert_eq!(sep_index_within(Some(3), 3), None);
+        assert_eq!(sep_index_within(None, 3), None);
+    }
+
+    #[test]
+    fn detect_row_mismatch() {
+        let rows = vec![
+            vec!["a".to_string(), "b".to_string()],
+            vec!["1".to_string(), "2".to_string()],
+        ];
+        assert!(!rows_mismatched(&rows, false));
+
+        let mismatch = vec![
+            vec!["a".to_string(), "b".to_string()],
+            vec!["1".to_string()],
+        ];
+        assert!(rows_mismatched(&mismatch, false));
+
+        let with_sep = vec![
+            vec!["a".to_string(), "b".to_string()],
+            vec!["---".to_string(), "---".to_string()],
+            vec!["1".to_string(), "2".to_string()],
+        ];
+        assert!(!rows_mismatched(&with_sep, false));
+
+        assert!(!rows_mismatched(&mismatch, true));
+    }
+}
